+++ app/src/main/java/com/sedsoftware/yaptalker/presentation/features/news/adapter/NewsAdapter.kt
@@ -41,9 +41,9 @@ class NewsAdapter @Inject constructor(
     }
   }
 
-  override fun onViewDetachedFromWindow(holder: ViewHolder?) {
+  override fun onViewDetachedFromWindow(holder: ViewHolder) {
     super.onViewDetachedFromWindow(holder)
-    holder?.itemView?.clearAnimation()
+    holder.itemView?.clearAnimation()
   }
 
   override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder =